複数ファイルの修正と機能追加
- `DirectX12.h` から `#include <Windows.h>` を削除
- `Framerate.cpp` に `#pragma comment(lib, "Winmm.lib")` を追加
- `WinSystem.h` で `#include <Windows.h>` の位置を変更し、`#include <stdint.h>` を追加
- `WinSystem.h` の `WindowProcedure` 関数の宣言を `__stdcall` に変更
- `DebugManager.cpp` の `ChangeFont` 関数で `fontcfg.MergeMode` を `0` から `1` に変更し、`io.Fonts->AddFontDefault()` を追加
- `ImGuiManager.cpp` の `Initialize` 関数で `ImGui_ImplWin32_Init` の呼び出し位置を変更
- `Logger.cpp` で `#include <debugapi.h>` を追加
- `Engine.vcxproj` の `PreprocessorDefinitions` に `_AMD64_` を追加
- `Engine.vcxproj` の `AdditionalDependencies` に `Network.lib` を追加
- `JsonParser.lib` のバイナリファイルが変更
- `json.h` の `Json::Tokenizer` クラスに `Initialize` 関数と `SetJsonString` 関数を追加
- `json.h` の `Json::Parser` クラスに `Initialize` 関数と `SetJsonString` 関数を追加
- `json.h` の `Json::Saver` クラスに `GetString` 関数を追加
- `Network.lib` のバイナリファイルが変更
- `TCP_IP.h` の `Receive` 関数の戻り値を `std::string` に変更
- `Timer.h` に `#include <Windows.h>` を追加
- `Input.h` で `#include <Windows.h>` の位置を変更
- `EmitterManager.h` で `ParseVector3`、`ParseVector4`、`ParseColor` 関数を `public` に移動
- `TransFadeInOut.cpp` の `Update` 関数で `Math::Lerp` の引数を `static_cast<float>` でキャスト

+++ Engine/Features/SceneTransition/TransFadeInOut.cpp
@@ -47,11 +47,11 @@ void TransFadeInOut::Update()
 
     if (countPhase_ == 0)
     {
-        opacity_ = Math::Lerp(0.0f, 1.0f, timer_->GetNow() / duration_);
+        opacity_ = Math::Lerp(0.0f, 1.0f, static_cast<float>(timer_->GetNow() / duration_));
     }
     else if (countPhase_ == 1)
     {
-        opacity_ = Math::Lerp(1.0f, 0.0f, timer_->GetNow() / duration_);
+        opacity_ = Math::Lerp(1.0f, 0.0f, static_cast<float>(timer_->GetNow() / duration_));
     }
 
     sprite_->SetColor(Vector4(0, 0, 0, opacity_));
